requirements.txt
django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
pytest-asyncio>=0.23.2
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
pytest-django>=4.7.0
django-background-tasks
django-cors-headers
djangorestframework-simplejwt
djangorestframework-simplejwt[django]
django-phonenumber-field
django-phonenumbers
